# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: XAI Blockchain API
  version: 1.1.0
  description: 'Production REST API for the XAI blockchain node.


    All endpoints require authenticated transport in production. Use the summaries
    and schemas in this document to integrate wallets, explorers, trading engines,
    and operational tooling.'
  contact:
    name: XAI Support
    url: https://xai-blockchain.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: http://localhost:5000
  description: Local development
- url: https://testnet-api.xai-blockchain.io
  description: Testnet
- url: https://api.xai-blockchain.io
  description: Production
...
paths:
...
  /airdrop/winners:
    get:
      tags:
      - Gamification
      summary: List airdrop winners
      description: Leaderboard of the most recent airdrop winners.
      parameters:
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  airdrops:
                    type: array
                    items:
                      type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /airdrop/user/{address}:
    get:
      tags:
      - Gamification
      summary: Get user airdrops
      description: Historical airdrop data for the provided address.
      parameters:
      - $ref: '#/components/parameters/AddressPathParam'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  address:
                    type: string
                  total_airdrops:
                    type: integer
                  history:
                    type: array
                    items:
                      type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /mining/streaks:
    get:
      tags:
      - Gamification
      summary: Mining streak leaderboard
      description: Returns the longest and most recent mining streaks.
      parameters:
      - $ref: '#/components/parameters/LimitParam'
      - name: sort_by
        in: query
        schema:
          type: string
          default: current_streak
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  leaderboard:
                    type: array
                    items:
                      type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /mining/streak/{address}:
    get:
      tags:
      - Gamification
      summary: Address mining streak
      description: Returns streak statistics for a specific miner.
      parameters:
      - $ref: '#/components/parameters/AddressPathParam'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  address:
                    type: string
                  stats:
                    type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /treasure/active:
    get:
      tags:
      - Gamification
      summary: List active treasures
      description: Returns all currently unclaimed treasure hunts.
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  count:
                    type: integer
                  treasures:
                    type: array
                    items:
                      type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /treasure/create:
    post:
      tags:
      - Gamification
      summary: Create treasure hunt
      description: Creates a new treasure hunt by locking a bounty behind a puzzle.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreasureCreateInput'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  treasure_id:
                    type: string
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /treasure/claim:
    post:
      tags:
      - Gamification
      summary: Claim treasure
      description: Validates the puzzle solution and rewards the claimer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreasureClaimInput'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  amount:
                    type: number
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /treasure/details/{treasure_id}:
    get:
      tags:
      - Gamification
      summary: Treasure details
      description: Returns the metadata for a treasure hunt.
      parameters:
      - $ref: '#/components/parameters/TreasureIdParam'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  treasure:
                    type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /timecapsule/pending:
    get:
      tags:
      - Gamification
      summary: Pending time capsules
      description: Lists time capsules that are still locked.
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  count:
                    type: integer
                  capsules:
                    type: array
                    items:
                      type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /timecapsule/{address}:
    get:
      tags:
      - Gamification
      summary: Address time capsules
      description: Returns sent and received time capsules for an address.
      parameters:
      - $ref: '#/components/parameters/AddressPathParam'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  address:
                    type: string
                  sent:
                    type: array
                  received:
                    type: array
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /refunds/stats:
    get:
      tags:
      - Gamification
      summary: Refund statistics
      description: Aggregated fee refund metrics.
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  stats:
                    type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /refunds/{address}:
    get:
      tags:
      - Gamification
      summary: Address refund history
      description: Lists the refunds that an address has received.
      parameters:
      - $ref: '#/components/parameters/AddressPathParam'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  address:
                    type: string
                  history:
                    type: array
                    items:
                      type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
...
components:
  parameters:
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
      description: Maximum number of records to return.
...
    AddressPathParam:
      name: address
      in: path
      required: true
      schema:
        type: string
...
    TreasureIdParam:
      name: treasure_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
        code:
          type: string
...
    TreasureCreateInput:
      type: object
      required:
      - creator
      - amount
      - puzzle_type
      - puzzle_data
      properties:
        creator:
          type: string
        amount:
          type: number
        puzzle_type:
          type: string
        puzzle_data:
          type: object
        hint:
          type: string
    TreasureClaimInput:
      type: object
      required:
      - treasure_id
      - claimer
      - solution
      properties:
        treasure_id:
          type: string
        claimer:
          type: string
        solution:
          type: string
...
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: X-API-Key
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- api_key: []
- bearer_auth: []